Never pass dots in prob vec functions that don't expose `estimator`
`estimator` is computed internally here, and is then passed on to `metric_vec_template()`. However, with `metric_tweak()` the signature of a prob metric has `estimator` in it. It should be completely ignored, but if dots were passed through then we would match `estimator` twice, once from the dots and once from setting it internally.