Fully commit to always swallowing `...`
This simplifies usage of `metric_set()` and `metric_tweak()` for various metrics that don't have an `estimator` argument, but still set it themselves internally and pass it on (like `accuracy()` and `classification_cost()`). Placing these metrics in `metric_tweak()` results in a function that has an `estimator` argument for completeness, which should be ignored, but can be problematic if we pass `...` through as we end up in a state where two `estimator` arguments are supplied, one from our internal usage and one from `metric_tweak()`.

The dots were never actually used, and were documented as not doing anything. So this shouldn't break anything.