Don't document `autoplot()` methods directly
These are registered dynamically, and that gives roxygen2 some issues. If you do `CTRL+SHIFT+D` they get documented like non-S3 functions, but if you do a `check()` in the console then they get documented as S3 methods. This is getting pretty annoying and noisy, so I'm just going to rely on the Examples section to show that these exist.